refactor: review comments
- Remove functions from `__init__` that are not exposed in Node
- Rename `Api` to `API` (Interface/Implementation)
- Move `UserContext` type alias to common types module
- Move error map util to common utils module
