+++ supertokens_python/recipe/webauthn/__init__.py
@@ -2,7 +2,6 @@ from typing import Optional
 
 from supertokens_python.recipe.webauthn.functions import (
     consume_recover_account_token,
-    create_new_recipe_user,
     create_recover_account_link,
     generate_recover_account_token,
     get_credential,
@@ -19,10 +18,9 @@ from supertokens_python.recipe.webauthn.functions import (
     sign_in,
     sign_in_options,
     sign_up,
-    update_user_email,
     verify_credentials,
 )
-from supertokens_python.recipe.webauthn.interfaces.api import ApiInterface, APIOptions
+from supertokens_python.recipe.webauthn.interfaces.api import APIInterface, APIOptions
 from supertokens_python.recipe.webauthn.interfaces.recipe import RecipeInterface
 from supertokens_python.recipe.webauthn.recipe import WebauthnRecipe
 from supertokens_python.recipe.webauthn.types.config import (
@@ -33,7 +31,7 @@ from supertokens_python.recipe.webauthn.types.config import (
 # Some Pydantic models need a rebuild to resolve ForwardRefs
 # Referencing imports here to prevent lint errors.
 # Caveat: These will be available for import from this module directly.
-ApiInterface  # type: ignore
+APIInterface  # type: ignore
 RecipeInterface  # type: ignore
 NormalisedWebauthnConfig  # type: ignore
 
@@ -51,7 +49,6 @@ __all__ = [
     "WebauthnConfig",
     "WebauthnRecipe",
     "consume_recover_account_token",
-    "create_new_recipe_user",
     "create_recover_account_link",
     "generate_recover_account_token",
     "get_credential",
@@ -68,6 +65,5 @@ __all__ = [
     "sign_in",
     "sign_in_options",
     "sign_up",
-    "update_user_email",
     "verify_credentials",
 ]
