+++ supertokens_python/recipe/webauthn/api/register_options.py
@@ -7,18 +7,18 @@ from supertokens_python.framework.response import BaseResponse
 from supertokens_python.recipe.webauthn.interfaces.recipe import (
     InvalidEmailErrorResponse,
 )
-from supertokens_python.recipe.webauthn.types.base import UserContext
+from supertokens_python.types.base import UserContext
 from supertokens_python.utils import send_200_response
 
 if TYPE_CHECKING:
     from supertokens_python.recipe.webauthn.interfaces.api import (
-        ApiInterface,
+        APIInterface,
         APIOptions,
     )
 
 
 async def register_options_api(
-    api_implementation: ApiInterface,
+    api_implementation: APIInterface,
     tenant_id: str,
     options: APIOptions,
     user_context: UserContext,
