+++ supertokens_python/recipe/webauthn/api/sign_in_options.py
@@ -3,18 +3,18 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Optional
 
 from supertokens_python.framework.response import BaseResponse
-from supertokens_python.recipe.webauthn.types.base import UserContext
+from supertokens_python.types.base import UserContext
 from supertokens_python.utils import send_200_response
 
 if TYPE_CHECKING:
     from supertokens_python.recipe.webauthn.interfaces.api import (
-        ApiInterface,
+        APIInterface,
         APIOptions,
     )
 
 
 async def sign_in_options_api(
-    api_implementation: ApiInterface,
+    api_implementation: APIInterface,
     tenant_id: str,
     options: APIOptions,
     user_context: UserContext,
