+++ supertokens_python/recipe/webauthn/types/base.py
@@ -1,12 +1,6 @@
-from typing import Any, Dict, List
-
-from typing_extensions import TypeAlias
+from typing import List
 
 from supertokens_python.types.response import CamelCaseBaseModel
-
-# TODO: Should `UserContext` be optional to handle `None` and init with `{}`?
-# TODO: Make this generic and re-use across codebase?
-UserContext: TypeAlias = Dict[str, Any]
 
 
 class WebauthnInfo(CamelCaseBaseModel):
